Doubled common code back into each module
For reasons unknown, having the common code in a separate module caused both recording modules to not have access to it (probably due to the sequence in which the modules are loaded), so the common code has been copied to each module for now. A fix is not necessary, but would be great.